cleanup older collection traits
