+++ src/par/collect_into/mod.rs
@@ -1,7 +1,4 @@
 mod fixed_vec;
 pub mod par_collect_into;
-pub mod par_fmap_fil_collect_into;
-pub mod par_map_collect_into;
-pub mod par_map_fil_collect_into;
 mod split_vec;
 mod vec;
